Use shared function for rendering markdown
Fixes #1740
