+++ client/browser/tslint.json
@@ -5,7 +5,6 @@
     "deprecation": {
       "severity": "warning"
     },
-    "import-blacklist": [true, "highlight.js"],
     "jsx-ban-props": false
   }
 }
