+++ shared/src/panel/views/PanelView.tsx
@@ -1,5 +1,4 @@
 import H from 'history'
-import marked from 'marked'
 import React from 'react'
 import { Observable } from 'rxjs'
 import { PanelViewWithComponent, ViewProviderRegistrationOptions } from '../../api/client/services/view'
@@ -8,6 +7,7 @@ import { Markdown } from '../../components/Markdown'
 import { ExtensionsControllerProps } from '../../extensions/controller'
 import { SettingsCascadeProps } from '../../settings/settings'
 import { createLinkClickHandler } from '../../util/linkClickHandler'
+import { renderMarkdown } from '../../util/markdown'
 import { EmptyPanelView } from './EmptyPanelView'
 import { HierarchicalLocationsView } from './HierarchicalLocationsView'
 
@@ -34,7 +34,7 @@ export class PanelView extends React.PureComponent<Props, State> {
             >
                 {this.props.panelView.content && (
                     <div className="px-2 pt-2">
-                        <Markdown dangerousInnerHTML={marked(this.props.panelView.content)} />
+                        <Markdown dangerousInnerHTML={renderMarkdown(this.props.panelView.content)} />
                     </div>
                 )}
                 {this.props.panelView.reactElement}
